Polishing the elaborate Indy path
This is coming along nicely.  I made a few changes that you can consider or reject:

1. Added leading comment with our names and what it does.
2. Adjusted the GMT logo so it also starts 2 cm from the side
3. Removed the PROJ_ELLIPSOID setting lke for the other script
4. Added more comments

Questions: The probling and trimming of the mp3 track: I assume you include this to show  how you need to get the same length and thus need original film length and then trim, Alternatively you could just use the final mp3 trimmed file as the original and avoid those two commands? Otherwise I think you need to add comments to explain this - or we do it whenwe discuss this script in the paper.